move backoff retry and jitter to `utils/misc`, fix imports
